Added token and admin access auth

+++ acl.service/src/main/java/controllers/GroupController.java
@@ -18,6 +18,7 @@ import org.springframework.web.servlet.ModelAndView;
 import org.springframework.http.HttpStatus;
 
 import authentication.EncryptionEngine;
+import authentication.TokenAuthentication;
 import dto.GroupDto;
 import model.Groups;
 import model.Resource;
@@ -32,17 +33,25 @@ public class GroupController {
 	GroupService groupService;
 	@Autowired
 	ResourceService resourceService;
-	
+    TokenAuthentication tokenauth = new TokenAuthentication();
+    String s="success";
+    String f="failed";
 	
 	//create new group by sending post req on /groups with the json data of group
 	@RequestMapping(method = RequestMethod.POST, consumes = "application/json")  
-    public void create(@RequestBody Groups group){
-    		groupService.createGroup(group.getgName(), group.getgDescription(), group.getgArbitraryAttributes(), group.getgResource(), group.getgUsers());
+    public String create(@RequestParam(value="token", required=true) String token, @RequestBody Groups group){
+		if(tokenauth.checkToken(token)) {
+			groupService.createGroup(group.getgName(), group.getgDescription(), group.getgArbitraryAttributes(), group.getgResource(), group.getgUsers());
+			return s;
+		}
+		return f;
     }
 	
 	//Return the group info + extra
     @RequestMapping(value = "/{gId}", method = RequestMethod.GET, produces = "application/json")  
-    public GroupDto group(@PathVariable int gId){
+    public GroupDto group(@RequestParam(value="token", required=true) String token, @PathVariable int gId){
+    	if(!tokenauth.checkToken(token))
+    		return null;
     	GroupDto groupDto = new GroupDto();
     	Groups group = groupService.getById(gId);
     	
@@ -55,17 +64,21 @@ public class GroupController {
     //update group
     @RequestMapping(value = "/{gId}", method = RequestMethod.PUT, consumes = "application/json")
     @ResponseStatus(HttpStatus.OK)
-    public void update(@PathVariable int gId,@RequestBody Groups group){
+    public String update(@RequestParam(value="token", required=true) String token, @PathVariable int gId,@RequestBody Groups group){
+    	if(!tokenauth.checkToken(token))
+    		return f;
     	groupService.updateGroup(gId, group.getgName(), group.getgDescription(), group.getgArbitraryAttributes(), group.getgResource());
+    	return s;
     }
     
     //delete group
     @RequestMapping(value = "/{gId}", method = RequestMethod.DELETE)  
-    public void delete(@PathVariable int gId){
+    public String delete(@RequestParam(value="token", required=true) String token, @PathVariable int gId){
+    	if(!tokenauth.checkToken(token))
+    		return f;
     	groupService.deleteGroup(gId);
+    	return s;
     }
-	
-    
     /*
 	
     @RequestMapping(value = "/addgroup", method = RequestMethod.POST, produces = "application/json")
